Initial Functionality Commit
The package is capable of connecting to an Apache Atlas endpoint, downloading and uploading entities and typedefs.

It supports a column lineage scaffolding that is similar to the Hive Bridge way of doing column lineage.

Entity and Relationship Type Defs are supported with the class objects.  Realtionships have a default table and column mapping
as per the hive bridge style (endDef1 is pointing to a table and creates the columns attribute, endDef2 is pointing
at the column and creates the table attribute).

Finally, the pyapacheatlas package can be installed after building the wheel files.