Adding work in progress for excel upload
Need to figure out how to do this more gracefully and take advantage of the client.

You should be able to just load a spreadsheet and it should find all the attributes.

You should be able to compose a combination of reading, validation, and then upload of the sheet.

You should be able to upload tables, column mappings, processes.

Those three things should be able to be validated against some sort of control process / set of typedefs

That control process is the hard part!