Adding Case Sensitivity to Reader
It turns out that the properties of an Atlas Entity are case sensitive.

Now the excel template reader does not attempt to standardize the column headers.
This has the downside that the Table, Column, Transformation, and Process columns
must be typed exactly as is.

However, now we get to have camelCasing in attributes work consistenly for
Tables, Columns, and EntityDefs tabs!