Support Non Lineage Entity Upload
Closes #29 by adding the readers.core.bulkEntities function that takes in
a json entity and transforms to an Atlas Entity.  Four columns are
required "typeName", "name", "qualifiedName", "classifications" and any
additional attributes / columns added are automatically added as
attributes to the entity.  If a cell is empty for a given row, it is
not added to the entity.  This way you can mix entity types in a single
sheet and not risk adding invalid attributes.

In addition in readers.excel an excel_bulkEntities function takes in a
sheet and parses it to entities.  It wraps the readers.core.bulkEntities.

A breaking change to older versions (pre 0.0b9) is introduced as the
defaults in the excel template now change to:
* ColumnsLineage from Columns
* TablesLineage from Tables
* BulkEntities is added for this feature
* EntityDefs is unchanged

This was necessary to help distinguish between the lineage section and
the bulk entity uploads that some users may want to do instead.

However, this doesn't solve the challenge of relationship attributes.
If you do want to upload a table and column entity (with no lineage)
this feature does not currently support that mapping.

Need to figure out an elegant way to distinguish between an attribute
and a relationship attribute (e.g. specifying the table entity for a
column).