Adding Relationship Attributes injection
Closes #31 and closes #34 by introducing the reserved column header of
"(Relationship) X" for the Bulk Entities tab. When a user provides
this as a column header, it will do a lookup against the previously
defined entities (order matters here) and then insert the minimum
json version of that entity. This is also enabled on the column
lineages tab but the limitation here is that there's no reason to use
it given that the relationship attributes are made for you.  Reserved
for future use.

You can now map the entities without having to use column lineages!