Get Entity by Name and Type and Simplify Apollo Sample (#51)
Atlas provides a way for you to reference the qualified name and type of the entity to pull its data down!

In addition, I realized that the Apollo sample was WAY over engineered when Atlas accepts the Entity Header instead of needing a GUID on either side of a relationship. That means you can push the work of finding the guid to the Atlas server and not rely on the search engine!