Support Batches of Bulk Uploads and Intelligently Batch Dependencies (#117)
When Atlas or Purview has a large set of entities to be uploaded (anecdotally, > 4,000 in Purview) the service may not be able to handle that many uploads. As a result, you may want to batch up your entities into groups that fit within the limit of your service.

However, the challenge is the number of dependencies when doing an upload. For example, a newly created table may have columns, schema, database, and server being created along side it. Those entities must be uploaded WITH the table otherwise the dummy guids will not be found during upload.

The batching will ignore grouping dependencies when they are NOT numbers.  This assumes a string or number (like -1 or "-1") is being used as a pointer to another entity in your upload being created or updated.