Feature/support full atlas entity (#133)
Added support for all the root attributes in an Atlas Entity from Atlas and Purview. This support requires using the new `AtlasUnInit` class which represents an uninitialized field. Since None / Null is a potentially valid value, by specifying the field is uninitialized, the `to_json` method will exclude returning an uninitialized field. However, it will return a None / Null field if it's set.  This is inconsistent with `EntityTypeDef`.

New methods were added on `AtlasEntity` to support adding custom and business attributes, adding classifications, and handling the `AtlasUnInit` initial state while supporting a sort of roll-back if there is an Exception.

In addition, I've added a `from_json` method to the `AtlasEntity` base class to support grabbing a single Atlas Entity's JSON and then converting it into an `AtlasEntity` (or `AtlasProcess`) object.

Deprecated methods on AtlasEntity and AtlasProcess were removed.