Updated the AtlasEntity.addRelationship method to support dealing with lists of AtlasEntities
Added a sample to demonstrate creating lineage WITHOUT a Process entity. This sample is of
an undocumented feature and should be treated as experimental and probably not take any
dependency on it but people keep asking about it / wanting it. It is used by the MSFT Purview
manual lineage feature.

Closes #263